fix: german translations (#370)
## Summary

Checking and improving the German localization. Follow-up to this:
https://bsky.app/profile/imlunahey.com/post/3m3jzf46fe22m

_Most_ of the changes are just for consistency and would have worked
fine, if maybe a bit awkward, in the original version. I'll go through
and leave a comment on the instances where the original might have
created problems.

One overarching thing is that it's inconsistently switching between the
informal "du" and the formal "Sie" to address the user, I've corrected
all of these to the informal form, which is much more common for user
interfaces these days. With LLMs it usually works well to include
something like "use informal language" in the prompt. DeepL even has an
option in their API for that iirc.

"Handle" wasn't translated in most instances and just kept as a loan
word. That works fine for a technical audience, but the wider population
will probably stumble over that, replacing with "Benutzername" (which is
just the translation for user name) would probably be appropriate.

---------

Signed-off-by: Robin <robin@roschlau.me>
Co-authored-by: Luna <105922175+ImLunaHey@users.noreply.github.com>